requirements.txt
colorama==0.4.6
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-nmap==0.7.1
scapy==2.6.1
